backend/infrastructure/gateway/requirements.txt
# RedFire API网关依赖

# 核心框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP客户端
httpx==0.25.2
aiohttp==3.9.1

# Redis支持
redis[hiredis]==5.0.1

# JWT认证
PyJWT==2.8.0
cryptography==41.0.8

# 监控和指标
prometheus-client==0.19.0

# 日志
structlog==23.2.0

# 配置管理
python-dotenv==1.0.0

# 测试依赖
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
httpx==0.25.2

# 开发工具
black==23.11.0
isort==5.12.0
mypy==1.7.1
flake8==6.1.0

# 文档生成
mkdocs==1.5.3
mkdocs-material==9.4.8
